Disable interpolation for configparser
Interpolation allows to replace values with this %(syntax).
Audiocontrol2 does not make use of this feature. However since some
secrets like the LastFM password are stored as-is in the config file,
the parser will crash when the secret contains the '%(' substring.
Disabling interpolation returns the secret as-is without trying to
interpolate it, allowing the use of such secrets.